+++ ShoppingCartModule/src/main/java/ecommerce/shoppingcartservice/service/CartLineService.java
@@ -1,13 +1,13 @@
 package ecommerce.shoppingcartservice.service;
 
-import ecommerce.shoppingcartservice.dto.CartLineRequest;
+import ecommerce.shoppingcartservice.model.dto.CartLineRequest;
 import ecommerce.shoppingcartservice.model.CartLine;
 
 public interface CartLineService {
 
-    void removeCartLine(int cartId);
+    void removeCartLine(Long cartId);
 
-    public boolean checkCartLineExistence(Integer cartId);
+    public boolean checkCartLineExistence(Long cartId);
 
     CartLine updateCartLine(CartLineRequest cartLineRequest);
 }
